move team4 to models.py within BettingSheets
