+++ BettingSheets/models.py
@@ -57,3 +57,34 @@ class UserGameSelection(models.Model):
            )
    high_risk = models.BooleanField(editable=False, max_length=10, default='False')
    selected_team =models.CharField(max_length=30, editable=True, blank=False, null=False);
+
+
+
+   #------ TEAM4 ----------
+   #-----------------------
+
+   def score(self):
+       if self.selected_team == self.game.winner(): #winner() function will reside in Game
+           return 1
+       else
+           return 0
+
+class Bettor(models.Model):
+    #def __init__(self, is_winston_cup, betting_sheet, weekly_points, winston_points, user_id):
+    is_winston_cup = models.BooleanField(default=False)
+    betting_sheet = models.ForeignKey('MasterBettingSheet', on_delete=models.CASCADE, null=True, verbose_name='Master Betting Sheet')
+    weekly_points = models.PositiveIntegerField(null=True, blank=False, default=None)
+    winston_points = models.PositiveIntegerField(null=True, blank=False, default=None)
+    user_id = models.PositiveIntegerField(null=True, blank=False, default=None)
+    currentPoints = models.PositiveIntegerField(null=True, blank=False, default=None)
+
+    def score_week(self, wk):
+        sheet = UserSheet.objects.filter(bettor = self, week = wk)
+        score = 0
+        for game in sheet.user_game_selection_set.all():
+            score = score + game.score()
+
+        return score
+
+    def score_winston(self):
+        #Do
